cleanup(team-org): drop playwright helpers and fixtures
